requirements.txt
# Ergeon internal PyPI server (for internal packages)
--extra-index-url=https://erg-bot:q8zgdmot3@pypi.ergeon.in/simple/

# Core dependencies for parcel-ai-json package
ultralytics>=8.0.0      # YOLOv8
pillow>=9.0.0           # Image processing
numpy>=1.20.0
torch>=2.0.0            # PyTorch for YOLO
torchvision>=0.15.0
pyproj>=3.0.0           # Geodesic coordinate transformations

# Tree detection with DeepForest
deepforest>=2.0.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Optional: Visualization (folium maps)
# folium>=0.14.0

# Development dependencies (install with: pip install -e ".[dev]")
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
